// src/app/common/model/old-data/old-product.interface.ts
export interface OldProductInterface {
	id?: number;
    sort?: number;
    name?: string;
    category?: string;
    showall?: number;
    colors?: string;
    colorhex?: string;
    height?: number;
    width?: number;
    length?: number;
    thickness?: number;
    sizetypes?: string;
    sizeslist?: string;
    size?: string;
    growth?: string;
    age?: string;
    gender?: number;
    fabric?: string;
    fulfillment?: string;
    description?: string;
    price?: number;
    quantity?: number;
    availability?: number;
    availabilityinfo?: string;
    novelty?: number;
    sale?: number;
    promotion?: number;
    thumb?: string;
    images?: string;
    published?: number;
}